Extract and search
Completed the notebook to cycle through the directory of downloaded json.gz files , extract each file in turn ,  search for key words on the fly and append the result to a pandas dataframe.
the process gets very slow as the dataframe becomes bigger.  My next step is to move this process to pyspark and save the contents directly  into cassandra